refactor: auth, member 테스트 패키지 분리
