+++ backend/src/test/java/hanglog/auth/domain/MemberFixture.java
@@ -1,4 +1,4 @@
-package hanglog.member.provider.fixture;
+package hanglog.auth.domain;
 
 public class MemberFixture {
